Results for this branch:

**Before (current `dev`)**

```
Warming up...                                                                   
OSVersion:              Microsoft Windows NT 6.2.9200.0                         
ProcessorCount:         16                                                      
ClockSpeed:             0 MHZ                                                   
Actor Count:            32                                                      
Messages sent/received: 30000000  (3e7)                                         
Thread count:           36                                                      
                                                                                
ActorBase    first start time:  9.07 ms                                         
ReceiveActor first start time: 31.74 ms                                         
                                                                                
Throughput, Msgs/sec, Start [ms], Total [ms],  Msgs/sec, Start [ms], Total [ms] 
         1,  7234000,     280.58,    4428.16,   7374000,     394.54,    4463.52 
        10, 25728000,       6.97,    1173.78,  24370000,       8.25,    1239.81 
        15, 31612000,     172.82,    1121.91,  28653000,     201.56,    1248.60 
        20, 32119000,       8.36,     942.50,  32223000,     456.02,    1387.07 
        30, 34168000,       6.98,     885.32,  37037000,     157.14,     967.64 
        40, 38860000,     422.94,    1195.30,  38910000,     374.58,    1146.58 
        50, 42492000,     108.18,     814.74,  37831000,       6.73,     800.57 
        60, 47770000,       6.38,     634.91,  43731000,     344.75,    1031.24 
        70, 42674000,      15.63,     719.58,  37313000,      29.59,     833.93 
        80, 48231000,       6.09,     628.83,  36991000,     439.48,    1250.90 
        90, 46656000,     234.63,     878.34,  45523000,     107.16,     766.39 
       100, 45592000,     311.36,     970.29,  40983000,     435.76,    1168.32 
       200, 47694000,     220.56,     850.49,  49916000,      31.44,     633.31 
       300, 50505000,       6.31,     600.64,  42253000,     220.19,     931.05 
       400, 49833000,       5.64,     608.58,  51457000,     204.24,     788.04 
```

**After (this branch)**

```
OSVersion:              Microsoft Windows NT 6.2.9200.0                         
ProcessorCount:         16                                                      
ClockSpeed:             0 MHZ                                                   
Actor Count:            32                                                      
Messages sent/received: 30000000  (3e7)                                         
Is Server GC:           True                                                    
Thread count:           39                                                      
                                                                                
ActorBase    first start time:  9.41 ms                                         
ReceiveActor first start time: 31.28 ms                                         
                                                                                
            ActorBase                          ReceiveActor                     
Throughput, Msgs/sec, Start [ms], Total [ms],  Msgs/sec, Start [ms], Total [ms] 
         1,  6720000,     385.29,    4849.47,   7070000,       8.98,    4252.94 
         5, 14866000,      10.72,    2029.43,  16198000,      76.93,    1929.07 
        10, 16882000,     313.47,    2090.91,  22222000,     358.21,    1708.76 
        15, 20746000,      93.91,    1540.11,  26501000,      75.01,    1207.64 
        20, 30060000,     125.67,    1123.90,  29940000,     357.96,    1360.31 
        30, 33482000,     202.39,    1099.26,  25817000,     265.89,    1428.08 
        40, 31746000,     202.12,    1147.79,  33557000,     157.07,    1051.96 
        50, 40268000,      13.76,     759.21,  30030000,       7.41,    1006.64 
        60, 24213000,     247.72,    1487.60,  40214000,      14.65,     761.10 
        70, 31578000,     216.86,    1166.99,  28037000,     139.38,    1210.09 
        80, 29732000,     281.68,    1291.06,  29910000,     155.59,    1159.32 
        90, 39840000,       6.10,     759.82,  29791000,     124.45,    1131.77 
       100, 32085000,     301.49,    1236.60,  35377000,     217.38,    1066.35 
       200, 38167000,     220.35,    1006.89,  37128000,     314.33,    1123.27 
       300, 33333000,     343.78,    1244.13,  28222000,     155.47,    1218.71 
       400, 37313000,     313.15,    1117.46,  36319000,      62.53,     889.36 
```

I suspect Gen 1/ Gen 2 garbage collection is increased here. Going to compare with another benchmark.