@Aaronontheweb 2021-04-23
Measuring actor memory allocation overhead for spawning a big tree of actors, which also involves a lot of messaging.

Running: https://github.com/akkadotnet/akka.net/blob/dev/src/benchmark/Akka.Benchmarks/Actor/SpawnActorBenchmarks.cs

**Before**

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19041.928 (2004/?/20H1)
AMD Ryzen 7 1700, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=5.0.201
  [Host]     : .NET Core 3.1.13 (CoreCLR 4.700.21.11102, CoreFX 4.700.21.11602), X64 RyuJIT
  Job-MJCVRM : .NET Core 3.1.13 (CoreCLR 4.700.21.11102, CoreFX 4.700.21.11602), X64 RyuJIT

InvocationCount=1  IterationCount=10  RunStrategy=Throughput  
UnrollFactor=1  WarmupCount=5  

```
|      Method | ActorCount |    Mean |    Error |   StdDev |       Gen 0 |      Gen 1 |     Gen 2 | Allocated |
|------------ |----------- |--------:|---------:|---------:|------------:|-----------:|----------:|----------:|
| Actor_spawn |     100000 | 2.336 s | 0.0737 s | 0.0487 s | 111000.0000 | 31000.0000 | 4000.0000 | 638.31 MB |

**After**

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19041.928 (2004/?/20H1)
AMD Ryzen 7 1700, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=5.0.201
  [Host]     : .NET Core 3.1.13 (CoreCLR 4.700.21.11102, CoreFX 4.700.21.11602), X64 RyuJIT
  Job-ADBXMO : .NET Core 3.1.13 (CoreCLR 4.700.21.11102, CoreFX 4.700.21.11602), X64 RyuJIT

InvocationCount=1  IterationCount=10  RunStrategy=Throughput  
UnrollFactor=1  WarmupCount=5  

```
|      Method | ActorCount |    Mean |    Error |   StdDev |      Gen 0 |      Gen 1 |     Gen 2 | Allocated |
|------------ |----------- |--------:|---------:|---------:|-----------:|-----------:|----------:|----------:|
| Actor_spawn |     100000 | 1.991 s | 0.0782 s | 0.0517 s | 96000.0000 | 24000.0000 | 4000.0000 | 541.77 MB |

Memory allocation is way down, actor spawn throughput is faster, Gen2 GC is the same, and Gen 0/1 actually decreased. Thought it might be GC pressure that is hurting throughput but apparently not....