@Aaronontheweb 2021-04-23
Re-running PingPong using Benchmark.NET

**Before**

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19041.928 (2004/?/20H1)
AMD Ryzen 7 1700, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=5.0.201
  [Host]     : .NET Core 3.1.13 (CoreCLR 4.700.21.11102, CoreFX 4.700.21.11602), X64 RyuJIT
  Job-YOVHXK : .NET Core 3.1.13 (CoreCLR 4.700.21.11102, CoreFX 4.700.21.11602), X64 RyuJIT

IterationCount=3  LaunchCount=3  RunStrategy=Monitoring  
WarmupCount=3  

```
|                                Method |     Mean |     Error |    StdDev |
|-------------------------------------- |---------:|----------:|----------:|
| Actor_ping_pong_single_pair_in_memory | 1.675 μs | 0.0393 μs | 0.0234 μs |

**After**

Operation failed. Couldn't complete the `Ask` operations on time. Failed repeatedly after multiple attempts at running it. I think this means that the queue structure isn't actually safe for writes and they're going missing.