@Aaronontheweb 2021-04-23
Yeah, I think we can write off this experiment - too complicated to make it work and outperform `ConcurrentQueue<T>` given our read / write semantics.