@Aaronontheweb 2021-04-28
Going to close this - as soon as you need to set two pointers inside a single operation, you really need a lock to do it properly.